#310 Refactor Home Navbar Hamburger Menu
