--- src/app/(home)/components/hamburgerMenu/index.ts
@@ -1,1 +1,0 @@
-export * from './hamburgerMenu'
